portfix

+++ libjl777/orders.h
@@ -1406,10 +1406,10 @@ int32_t init_rambooks(char *base,char *rel,uint64_t baseid,uint64_t relid)
 int32_t iQ_exchangestr(char *exchange,struct InstantDEX_quote *iQ)
 {
     if ( iQ->baseiQ != 0 && iQ->reliQ != 0 )
-        sprintf(exchange,"%s_%s",Exchanges[iQ->baseiQ->exchangeid].name,Exchanges[iQ->reliQ->exchangeid].name);
+        sprintf(exchange,"%s_%s",Exchanges[(int32_t)iQ->baseiQ->exchangeid].name,Exchanges[(int32_t)iQ->reliQ->exchangeid].name);
     else if ( iQ->exchangeid == INSTANTDEX_NXTAEID && iQ->timestamp > time(NULL) )
         strcpy(exchange,"unconf");
-    else strcpy(exchange,Exchanges[iQ->exchangeid].name);
+    else strcpy(exchange,Exchanges[(int32_t)iQ->exchangeid].name);
     return(0);
 }
 
@@ -1417,7 +1417,7 @@ cJSON *makeoffer_legjson(char *baserelstr,struct InstantDEX_quote *iQ,char *fiel
 {
     char numstr[64];
     cJSON *obj = cJSON_CreateObject();
-    cJSON_AddItemToObject(obj,baserelstr,cJSON_CreateString(Exchanges[iQ->exchangeid].name));
+    cJSON_AddItemToObject(obj,baserelstr,cJSON_CreateString(Exchanges[(int32_t)iQ->exchangeid].name));
     if ( iQ->nxt64bits != 0 )
         sprintf(numstr,"%llu",(long long)iQ->nxt64bits), cJSON_AddItemToObject(obj,"NXT",cJSON_CreateString(numstr));
     sprintf(numstr,"%llu",(long long)calc_quoteid(iQ)), cJSON_AddItemToObject(obj,"quoteid",cJSON_CreateString(numstr));
